requirements.txt
# Dépendances Python pour le projet SONAGED

# Base de données
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Traitement des données
pandas>=2.0.0
numpy>=1.24.0

# Visualisation Streamlit
streamlit>=1.28.0
plotly>=5.17.0
folium>=0.14.0
streamlit-folium>=0.15.0

# Utilitaires
python-dotenv>=1.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
matplotlib>=3.7.0

# PostGIS (si nécessaire pour certaines opérations)
# geoalchemy2>=0.14.0
